The big commit. The source compiles and links in mingw now. (and somewhat runs, but not really because lots of msvc assembler blitting routines are commented out). It also compiles (but not links) on unix. Changes:
-Added very basic makefile
-removed #include "Stdafx.h" from all source files, gcc doesn't have precompiled headers like msvc.
-Added some missing include in several files
-Removed strange double-include protection from codewiz.hpp
-Removed DOS-compilation stuff from codewiz.hpp
-Renamed TRUE to true, FALSE to false, BOOL to bool and removed the definitions
 for them from codewiz.hpp
-Removed BYTE,WORD,DWORD definitions from codewiz.hpp these are already defined
 in Types/LibTypes.hpp
-Commented memset assembler routines in codewiz.hpp
-removed special WATCOM/MSVC #pragmas in lots of places in codewiz.hpp
-changed type of PALETTE_LENGTH to size_t
-added a SWAP function to codewiz.hpp, couldn't find one elsewhere in the
 project. Also added a SWAP function with an explicit temporary variable which
 is used at some places in the project (does this bring speed?)
-made destructor of Surface, Component, SpritePacked virtual
-changed usage of zeroMem to memset
-use memcpy instead of memcpyB
-Added a stub of RDTSC function to cTimeStamp.cpp, need an implementation there.-Added a dummy GameTemplateView.hpp. Needs investigation why the Lib includes
 files from the main game.
-copied KeyScanCodeDefs.h from main game
-removed #if _MSC_VER > 1000 #pragma once #endif randomly from several files...
-removed /= (const float&) in fXYZ class (/= (float) is enough)
-gave up with the Lib/Main Game dependencies and added include paths to the maingame in the Makefiles...
-removed "aliasdef.h" from several files (it contained yet another definition of  a boolean type...), several renames again :-/
-added int return type to getUnitInfoDrawLayer in GameConfig
-changed iRect constructor to use const types
-disabled all joystick code
-casted several floats to ints in ColorTable.cpp to avoid warnings
-Commented out several uses of _splitpath (which is win32 only) at FileList.cpp
-fixed some upper/lowercase errors in #include statements
-commented out code that uses _findfirst in Surface.cpp and PackedSurface.cpp,
 this is a win32 only function
-casted some floats to int in Palette.cpp to avoid warnings
-commented out several assembler routines in Span.cpp
-removed custom memcpy from Surface.cpp
-changed a MessageBox to fprintf(stderr, ...) in Surface.cpp
-added implementation of MIN and MAX to codewiz.hpp
-renamed stricmp to strcasecmp in UtilInterface.cpp
-made some functions in SpritePacked and Sprite const-correct
-commented out lots of code in DirtKickParticle2D, because a variable didn't
seem to be defined. Did teh same in GroundExplosionParticle
-some more const-correctness fixes in GroundExplosionParticle and
TrajectoryParticle.
-made some functions in WorldViewInterface const-correct, same in ViewCamera
-commented out some direct sound code in ParticleInterface.cpp
-renamed some occurences of NULL to 0 in ParticleSystem2D
-commented out isRolloverEnabled in AbstractButton.hpp, this variable doesn't
seem to exist.
-removed Lib/loadSaveData.* it wasn't used (some joystick stuff, we wanna remove        the joystick anyway).

+++ trunk/netpanzer/src/Lib/2D/RGBColor.cpp
@@ -15,7 +15,7 @@ You should have received a copy of the GNU General Public License
 along with this program; if not, write to the Free Software
 Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 */
-#include "stdafx.hpp"
+
 #include "RGBColor.hpp"
 
 
@@ -32,7 +32,7 @@ RGBColor::RGBColor(BYTE red, BYTE green, BYTE blue)
 //--------------------------------------------------------------------------
 int RGBColor::getBrightnessInt()
 {
-	return 255.0f / 768.0f * float(red + blue + green);
+	return (int) (255.0f / 768.0f * float(red + blue + green));
 }
 
 // getBrightnessFloat
@@ -41,5 +41,5 @@ int RGBColor::getBrightnessInt()
 //--------------------------------------------------------------------------
 int RGBColor::getBrightnessFloat()
 {
-	return float(red + blue + green) / 768.0f;
+	return (int) (float(red + blue + green) / 768.0f);
 }
